Move the tests for WASI args to a wasi's subdirectory
